Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/clients-matchall-include-uncontrolled.https-expected.txt



Harness Error (FAIL), message = Error: assert_array_equals: property 1, expected false but got true

TIMEOUT Verify matchAll() with windows respect includeUncontrolled Test timed out
NOTRUN Verify matchAll() with shared workers respect includeUncontrolled
